Refactor package structure to align with extension naming conventions and improve organization
