--- src/main/java/com/mule/whisperer/api/STTParamsModelDetails.java
@@ -1,46 +1,0 @@
-package com.mule.whisperer.api;
-import com.mule.whisperer.internal.metadata.ModelNameProvider;
-import org.mule.runtime.api.meta.ExpressionSupport;
-import org.mule.runtime.extension.api.annotation.Expression;
-import org.mule.runtime.extension.api.annotation.param.Optional;
-import org.mule.runtime.extension.api.annotation.param.Parameter;
-import org.mule.runtime.extension.api.annotation.values.OfValues;
-
-public class STTParamsModelDetails {
-	@Parameter
-	@Expression(ExpressionSupport.SUPPORTED)
-	@OfValues(ModelNameProvider.class)
-	@Optional(defaultValue = "whisper-1")
-	private String modelName;
-
-	public String getModelName() {
-		return modelName;
-	}
-
-	@Parameter
-	@Expression(ExpressionSupport.SUPPORTED)
-	@Optional(defaultValue = "en")
-	private String language;
-
-	public String getLanguage() {	
-		return language;
-	}
-
-	@Parameter
-	@Expression(ExpressionSupport.SUPPORTED)
-	@Optional(defaultValue = "0.9")
-	private Number temperature;
-
-	public Number getTemperature() {
-		return temperature;
-	}
-
-	@Parameter
-	@Expression(ExpressionSupport.SUPPORTED)
-	@Optional(defaultValue = "false")
-	private boolean verbose;
-
-	public boolean isVerbose() {
-		return verbose;
-	}
-}
