--- src/main/java/com/mule/whisperer/api/error/ConnectorError.java
@@ -1,10 +1,0 @@
-package com.mule.whisperer.api.error;
-
-import org.mule.runtime.extension.api.error.ErrorTypeDefinition;
-
-public enum ConnectorError implements ErrorTypeDefinition<ConnectorError> {
-    CONNECTION_INCOMPATIBLE,
-    TRANSCRIPTION,
-    GENERATION,
-    TIMEOUT
-}
