--- src/main/java/com/mule/whisperer/internal/error/TranscriptionErrorTypeProvider.java
@@ -1,18 +1,0 @@
-package com.mule.whisperer.internal.error;
-
-import com.mule.whisperer.api.error.ConnectorError;
-import org.mule.runtime.extension.api.annotation.error.ErrorTypeProvider;
-import org.mule.runtime.extension.api.error.ErrorTypeDefinition;
-
-import java.util.HashSet;
-import java.util.Set;
-
-public class TranscriptionErrorTypeProvider implements ErrorTypeProvider {
-    @Override
-    public Set<ErrorTypeDefinition> getErrorTypes() {
-        Set<ErrorTypeDefinition> errorTypes = new HashSet<>();
-        errorTypes.add(ConnectorError.TRANSCRIPTION);
-        errorTypes.add(ConnectorError.TIMEOUT);
-        return errorTypes;
-    }
-}
