+++ src/main/java/org/mule/extension/whisperer/api/OpenAiTranscriptionAttributes.java
@@ -1,0 +1,22 @@
+package org.mule.extension.whisperer.api;
+
+public class OpenAiTranscriptionAttributes {
+    private String language;
+    private Double duration;
+
+    public String getLanguage() {
+        return language;
+    }
+
+    public void setLanguage(String language) {
+        this.language = language;
+    }
+
+    public Double getDuration() {
+        return duration;
+    }
+
+    public void setDuration(Double duration) {
+        this.duration = duration;
+    }
+}
