+++ src/main/java/org/mule/extension/whisperer/api/error/ConnectorError.java
@@ -1,0 +1,10 @@
+package org.mule.extension.whisperer.api.error;
+
+import org.mule.runtime.extension.api.error.ErrorTypeDefinition;
+
+public enum ConnectorError implements ErrorTypeDefinition<ConnectorError> {
+    CONNECTION_INCOMPATIBLE,
+    TRANSCRIPTION,
+    GENERATION,
+    TIMEOUT
+}
